fix: move chakraTheme to client root
